requirements.txt
yfpy>=10.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
